% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw04pDmando-package.R
\docType{package}
\name{hw04pDmando-package}
\alias{hw04pDmando}
\alias{hw04pDmando-package}
\title{hw04pDmando: Two Functions for HW4}
\description{
If this package works it will help me get an A in this class and it has two key functions.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
