% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myseq_n.R
\name{myseq_n}
\alias{myseq_n}
\title{Recursive function}
\usage{
myseq_n(x, n)
}
\arguments{
\item{x}{numeric}

\item{n}{numeric}
}
\value{
the solution to the recursive function.
}
\description{
This function creates and solves a recursive function that requires three x variables inputted.
}
\examples{
myseq_n(x = c(2, 4, 3), n = 3)
}
