% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myseqplot.R
\name{myseqplot}
\alias{myseqplot}
\title{Plots recursive function}
\usage{
myseqplot(inputdata)
}
\arguments{
\item{inputdata}{dataset}
}
\value{
Plot of the recursive function output.
}
\description{
This function plots the output of a recursive function on a data set.
}
\examples{
my_data <- tibble::tribble(
~x, ~y, ~z, ~n,
2,4,3,3,
2,4,3,4,
2,4,3,5,
2,4,3,6,
2,4,3,7,
2,4,3,8,
2,4,3,9,
2,4,3,10,
2,4,3,12)
myseqplot(my_data)
}
