% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myseqplot.R
\name{myseqplot}
\alias{myseqplot}
\title{Plots recursive function}
\usage{
myseqplot(inputdata)
}
\arguments{
\item{inputdata}{dataset}
}
\value{
Plot of the recursive function output.
}
\description{
This function plots the output of a recursive function on a data set.
}
\examples{
myseqplot(inputdata)
}
